refactor(proxy): split internal proxy modules
