+++ Sources/XcodeMCPProxy/XcodeMCPProxyExports.swift
@@ -1,0 +1,1 @@
+@_exported import XcodeMCPProxyCore
