+++ Sources/XcodeMCPProxyCore/Discovery.swift
