+++ Sources/XcodeMCPProxyCore/ProxyConfig.swift
