+++ Sources/XcodeMCPProxyCore/ProxyLogging.swift
