+++ Sources/XcodeMCPProxyTransportHTTP/HTTPRequestValidator.swift
